Update sway script to not break on broken symlink
Using Mod+Shift+Return starts a terminal with the same CWD of the
application contained in the focused window. If that application does
not have a CWD (such as a browser), an error occurs. This is caused
by readlink reporting success when it reads a deleted link,
which bypasses the "error handling". Therefore:
- Add -f option to readlink to cause it to fail on deleted links